feat(client): Implemented JOIN and LEAV commands
- Moved player related stuff to common
- Implemented structure for holding players
- Implemented a first draft of the JOIN and LEAV commands
- Moved further stuff to common
- Some refactoring concerning error/info codes

+++ mazegame-server-ktor/src/main/kotlin/de/dreamcube/mazegame/server/control/server_information.kt
@@ -35,7 +35,7 @@ data class PlayerInformationDto(
     val delayOffset: Long,
     val totalPlayTime: PlayTimeDto,
     val currentPlayTime: PlayTimeDto,
-    val currentPointsPerMinute: Int,
+    val currentPointsPerMinute: Double,
     val spectator: Boolean
 )
